- Add `ScrollViewer` control with visual cues for smoother user experience. - Update and unify namespace structure for clarity across files. - Modify `ShellView` and `Sidebar` templates to improve customization and accessibility (e.g., dynamic pane size, header font weight). - Replace ambiguous icon-related properties in `CardView` and improve related bindings. - Fix issues in resource organization (e.g., moved `ScrollViewer` styling resources, corrected missing/newlines, enhanced button style previews). - Minor changes: organize imports, update design preview definitions, adjust spacing converters.
