+++ samples/TestApp/TestApp/Samples/Misc/GradualGraphView.axaml
@@ -3,7 +3,6 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:g="clr-namespace:Zafiro.Avalonia.DataViz.Graphs.Control;assembly=Zafiro.Avalonia.DataViz"
-             xmlns:vm="clr-namespace:TestApp.Samples.Controls"
              xmlns:misc="clr-namespace:TestApp.Samples.Misc"
              mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
              x:Class="TestApp.Samples.Misc.GradualGraphView"
@@ -12,7 +11,7 @@
         <Button ZIndex="1" DockPanel.Dock="Top" Content="Load" Command="{Binding GradualGraph.Load}" />
         <g:GraphControl Margin="-1000" Graph="{Binding GradualGraph}">
             <g:GraphControl.ItemTemplate>
-                <DataTemplate DataType="vm:Person">
+                <DataTemplate DataType="misc:Person">
                     <Ellipse Height="{Binding $self.Width}" Fill="Red">
                         <Ellipse.Width>
                             <MultiBinding Converter="{x:Static misc:GraphConverters.Importance}">
@@ -25,7 +24,7 @@
                 </DataTemplate>
             </g:GraphControl.ItemTemplate>
             <g:GraphControl.EdgeTemplate>
-                <DataTemplate DataType="vm:Friendship">
+                <DataTemplate DataType="misc:Friendship">
                     <Line Stroke="Black" StrokeThickness="3">
                         <Line.StartPoint>
                             <MultiBinding Converter="{x:Static g:Converters.CoordinatesToPoint}">
