+++ samples/TestApp/TestApp/Samples/Misc/GradualGraphViewModel.cs
@@ -5,7 +5,6 @@ using System.Reactive;
 using System.Reactive.Linq;
 using Avalonia.Controls.Selection;
 using ReactiveUI;
-using TestApp.Samples.Controls;
 using Zafiro.Avalonia.DataViz.Graphs.Control;
 using Zafiro.DataAnalysis.Clustering;
 using Zafiro.DataAnalysis.Graphs;
