+++ src/Zafiro.Avalonia/Controls/Shell/ShellView.axaml.cs
@@ -1,16 +1,25 @@
 using Avalonia.Controls.Primitives;
 using Zafiro.UI.Shell;
 
-namespace Zafiro.Avalonia.Shell;
+namespace Zafiro.Avalonia.Controls.Shell;
 
 public class ShellView : TemplatedControl
 {
     public static readonly StyledProperty<IShell> ShellProperty = AvaloniaProperty.Register<ShellView, IShell>(
         nameof(Shell));
 
+    public static readonly StyledProperty<double> OpenPaneLengthProperty = AvaloniaProperty.Register<ShellView, double>(
+        nameof(OpenPaneLength));
+
     public IShell Shell
     {
         get => GetValue(ShellProperty);
         set => SetValue(ShellProperty, value);
+    }
+
+    public double OpenPaneLength
+    {
+        get => GetValue(OpenPaneLengthProperty);
+        set => SetValue(OpenPaneLengthProperty, value);
     }
 }
