+++ src/Zafiro.Avalonia/Styles/Button.axaml
@@ -1,8 +1,9 @@
 ﻿<Styles xmlns="https://github.com/avaloniaui"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
 
-    
+
     <Styles.Resources>
+
         <ControlTheme x:Key="TransparentButton" TargetType="Button" BasedOn="{StaticResource {x:Type Button}}">
             <Setter Property="Background" Value="Transparent" />
             <Setter Property="Margin" Value="1" />
@@ -20,6 +21,7 @@
                 <Setter Property="BorderBrush" Value="{DynamicResource SoftBorder}" />
             </Style>
         </ControlTheme>
+
         <ControlTheme x:Key="FullTransparentButton" TargetType="Button" BasedOn="{StaticResource {x:Type Button}}">
             <Setter Property="Background" Value="Transparent" />
             <Setter Property="Margin" Value="1" />
@@ -53,6 +55,7 @@
             </Style>
             <Style Selector="^:checked /template/ ContentPresenter#PART_ContentPresenter">
                 <Setter Property="Background" Value="Transparent" />
+                <Setter Property="Foreground" Value="{DynamicResource ToggleButtonForeground}" />
             </Style>
             <Style Selector="^:checked:pointerover /template/ ContentPresenter#PART_ContentPresenter">
                 <Setter Property="Background" Value="Transparent" />
@@ -61,9 +64,12 @@
     </Styles.Resources>
 
     <Design.PreviewWith>
-        <Border Padding="20">
-            <Button IsEnabled="True" Content="Transparent" Theme="{StaticResource TransparentButton}" />
-        </Border>
+        <StackPanel>
+            <Border Padding="20">
+                <Button IsEnabled="True" Content="Transparent" Theme="{StaticResource TransparentButton}" />
+            </Border>
+            <RepeatButton IsEnabled="True" Content="Transparent" Theme="{StaticResource TransparentButton}" />
+        </StackPanel>
     </Design.PreviewWith>
 
 </Styles>
