+++ src/Zafiro.Avalonia/Zafiro.Avalonia.csproj
@@ -18,27 +18,10 @@
         <PackageReference Include="Projektanker.Icons.Avalonia"/>
         <PackageReference Include="System.Linq.Async"/>
         <PackageReference Include="Zafiro"/>
-<!--        <PackageReference Include="Zafiro.UI"/>-->
+        <PackageReference Include="Zafiro.UI"/>
         <PackageReference Include="ReactiveUI.SourceGenerators">
             <PrivateAssets>all</PrivateAssets>
             <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
         </PackageReference>
-    </ItemGroup>
-    <ItemGroup>
-        <Compile Update="Controls\Shell\ShellView.axaml.cs">
-            <DependentUpon>ShellView.axaml</DependentUpon>
-            <SubType>Code</SubType>
-        </Compile>
-        <Compile Update="Controls\Shell\Sidebar.axaml.cs">
-            <DependentUpon>Sidebar.axaml</DependentUpon>
-            <SubType>Code</SubType>
-        </Compile>
-        <Compile Update="Controls\BulletList.axaml.cs">
-            <DependentUpon>List.axaml</DependentUpon>
-            <SubType>Code</SubType>
-        </Compile>
-    </ItemGroup>
-    <ItemGroup>
-      <ProjectReference Include="..\..\..\Zafiro\src\Zafiro.UI\Zafiro.UI.csproj" />
     </ItemGroup>
 </Project>
